reuse map component for workspaces list
