--- applications/fishing-map/src/features/map-workspaces-list/MapWorkspacesList.module.css
@@ -1,3 +1,0 @@
-.container {
-  flex: 1;
-}
